# Type Inference & Performance Improvements