Improve performance:
1. Restore type information to `MapCallback` and `ObserverFunction`
2. Allow method specializations for functions which are likely to be called many times
3. Eliminate unnecessary `invokelatest` 

With Julia 1.9's native code caching, some compile-time optimizations have become counterproductive.

Before this PR:
```julia
julia> using BenchmarkTools
       using Observables: Observable, @map
       const a = Observable(1.0);  @btime notify(a)
       const b = @map &a+1;        @btime notify(a)
       const c = @map &a+2;        @btime notify(a)
       a[], b[], c[]
  5.100 ns (0 allocations: 0 bytes)
  575.275 ns (10 allocations: 160 bytes)
  1.210 μs (20 allocations: 320 bytes)
(1.0, 2.0, 3.0)
```

After this PR:
```julia
julia> using BenchmarkTools
       using Observables: Observable, @map
       const a = Observable(1.0);  @btime notify(a)
       const b = @map &a+1;        @btime notify(a)
       const c = @map &a+2;        @btime notify(a)
       a[], b[], c[]
  5.200 ns (0 allocations: 0 bytes)
  30.723 ns (1 allocation: 16 bytes)
  55.193 ns (2 allocations: 32 bytes)
(1.0, 2.0, 3.0)
```

Because `invokelatest` is still used in `notify`, it is not necessary in the functor for `MapCallback`. Thus this PR does not affect issue #50, which remains closed.